requirements.txt
flask
python-dotenv
pymongo[srv]
gunicorn
numpy
scikit-learn
